handle utilisateur delete and status controe fortickets

+++ src/main/java/com/csys/template/service/TicketService.java
@@ -9,11 +9,18 @@ import com.csys.template.dtoResponse.TicketResponseDTO;
 import com.csys.template.factory.TicketFactory;
 import com.csys.template.repository.TicketRepository;
 import com.csys.template.util.WhereClauseBuilder;
+
+import liquibase.pro.packaged.he;
+import liquibase.pro.packaged.js;
+
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import com.csys.template.util.Helper;;
 
 @Service
 @Transactional
@@ -36,19 +43,8 @@ public class TicketService {
         log.debug("Request to update Ticket: {}", ticketId);
         Ticket existingTicket = ticketRepository.findById(ticketId)
                 .orElseThrow(() -> new IllegalArgumentException("ticket.NotFound"));
-        
-        // Create a temporary entity from the DTO to get related objects
-        Ticket updatedData = TicketFactory.toEntity(ticketRequestDTO);
-        
-        existingTicket.setTitre(updatedData.getTitre());
-        existingTicket.setDescription(updatedData.getDescription());
-        existingTicket.setPriorite(updatedData.getPriorite());
-        existingTicket.setStatue(updatedData.getStatue());
-        existingTicket.setParentTicket(updatedData.getParentTicket());
-        existingTicket.setIdClient(updatedData.getIdClient());
-        existingTicket.setModule(updatedData.getModule());
-        existingTicket.setIdUtilisateur(updatedData.getIdUtilisateur());
 
+        TicketFactory.updateFromDTO(existingTicket, ticketRequestDTO);
         ticketRepository.save(existingTicket);
         return TicketFactory.toResponseDTO(existingTicket);
     }
@@ -61,20 +57,37 @@ public class TicketService {
     }
 
     @Transactional(readOnly = true)
-    public List<TicketResponseDTO> findAll(Status statue, Integer idModule, Priorite priorite) {
+    public List<TicketResponseDTO> findAll(Status statue, Integer idModule, Priorite priorite,Boolean[] actifs) {
         log.debug("Request to get All Tickets with filters");
         QTicket qTicket = QTicket.ticket;
         WhereClauseBuilder builder = new WhereClauseBuilder()
                 .optionalAnd(statue, () -> qTicket.statue.eq(statue))
                 .optionalAnd(idModule, () -> qTicket.module().id.eq(idModule))
+                .optionalAnd(actifs, () -> qTicket.actif.in(actifs))
                 .optionalAnd(priorite, () -> qTicket.priorite.eq(priorite));
 
         List<Ticket> result = (List<Ticket>) ticketRepository.findAll(builder);
         return TicketFactory.toResponseDTOs(result);
     }
 
-    public void delete(Integer id) {
+    public ResponseEntity<?> delete(Integer id) {
+        Ticket ticket = ticketRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("ticket.NotFound"));
+        if(ticket.getModule()!=null){
+            return ResponseEntity.status(HttpStatus.LOCKED)
+                    .body(java.util.Map.of("message", "ticket " + ticket.getId() + " cannot Delete With Module"));
+        }
+        if (ticket.getIdUtilisateur() != null) {
+            return ResponseEntity.status(HttpStatus.LOCKED)
+                    .body(java.util.Map.of("message", "ticket " + ticket.getId() + " cannot Delete With User"));
+        }
+        TicketResponseDTO ticketResponseDTO = TicketFactory.toResponseDTO(ticket);
+        if (!ticketResponseDTO.getChildTickets().isEmpty()) {
+            return ResponseEntity.status(HttpStatus.LOCKED)
+                    .body(java.util.Map.of("message", "ticket " + ticketResponseDTO.getId() + " cannot Delete With Child Tickets"));
+        }
         log.debug("Request to delete Ticket: {}", id);
         ticketRepository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 }
